refactor: move all exceptions to one module
