+++ playa/__init__.py
@@ -6,7 +6,7 @@ Basic usage:
     pdf = playa.open(path)
 """
 
-import io
+import builtins
 from contextlib import contextmanager
 from os import PathLike
 
@@ -19,6 +19,6 @@ __version__ = "0.0.1"
 @contextmanager
 def open(path: PathLike, password: str = "") -> PDFDocument:
     """Open a PDF document from a path on the filesystem."""
-    with io.open(path, "rb") as infh:
+    with builtins.open(path, "rb") as infh:
         parser = PDFParser(infh)
         yield PDFDocument(parser, password)
