+++ playa/image.py
@@ -10,6 +10,7 @@ except ImportError:
     # Literal was introduced in Python 3.8
     from typing_extensions import Literal  # type: ignore[assignment]
 
+from playa.exceptions import PDFValueError
 from playa.jbig2 import JBIG2StreamReader, JBIG2StreamWriter
 from playa.layout import LTImage
 from playa.pdfcolor import (
@@ -19,7 +20,6 @@ from playa.pdfcolor import (
     LITERAL_INLINE_DEVICE_GRAY,
     LITERAL_INLINE_DEVICE_RGB,
 )
-from playa.pdfexceptions import PDFValueError
 from playa.pdftypes import (
     LITERALS_DCT_DECODE,
     LITERALS_FLATE_DECODE,
