+++ playa/layout.py
@@ -15,8 +15,8 @@ from typing import (
     cast,
 )
 
+from playa.exceptions import PDFTypeError, PDFValueError
 from playa.pdfcolor import PDFColorSpace
-from playa.pdfexceptions import PDFTypeError, PDFValueError
 from playa.pdffont import PDFFont
 from playa.pdfinterp import Color, PDFGraphicState
 from playa.pdftypes import PDFStream
@@ -106,7 +106,7 @@ class LAParams:
                 "LAParam boxes_flow should be None, or a number between -1 and +1"
             )
             if not (
-                isinstance(self.boxes_flow, int) or isinstance(self.boxes_flow, float)
+                isinstance(self.boxes_flow, (float, int))
             ):
                 raise PDFTypeError(boxes_flow_err_msg)
             if not -1 <= self.boxes_flow <= 1:
