+++ playa/pdfpage.py
@@ -3,12 +3,12 @@ import logging
 from typing import Any, BinaryIO, Container, Dict, Iterator, List, Optional, Set, Tuple
 
 from playa import settings
+from playa.exceptions import PDFObjectNotFound, PDFValueError
 from playa.pdfdocument import (
     PDFDocument,
     PDFNoPageLabels,
     PDFTextExtractionNotAllowed,
 )
-from playa.pdfexceptions import PDFObjectNotFound, PDFValueError
 from playa.pdfparser import PDFParser
 from playa.pdftypes import dict_value, int_value, list_value, resolve1
 from playa.psparser import LIT
