--- playa/psexceptions.py
@@ -1,18 +1,0 @@
-class PSException(Exception):
-    pass
-
-
-class PSEOF(PSException):
-    pass
-
-
-class PSSyntaxError(PSException):
-    pass
-
-
-class PSTypeError(PSException):
-    pass
-
-
-class PSValueError(PSException):
-    pass
